feat: create routes

+++ DotNetApi/GoogleMaps/GoogleMapsPlaceResponse.cs
@@ -17,7 +17,7 @@ public class Candidate
   public string Name { get; set; }
   [JsonProperty("place_id")]
   public string PlaceId { get; set; }
-  public double Rating { get; set; }
+  public float Rating { get; set; }
 }
 
 public class Geometry
@@ -28,8 +28,8 @@ public class Geometry
 
 public class Location
 {
-  public double Lat { get; set; }
-  public double Lng { get; set; }
+  public float Lat { get; set; }
+  public float Lng { get; set; }
 }
 
 public class Viewport
@@ -40,12 +40,12 @@ public class Viewport
 
 public class Northeast
 {
-  public double Lat { get; set; }
-  public double Lng { get; set; }
+  public float Lat { get; set; }
+  public float Lng { get; set; }
 }
 
 public class Southwest
 {
-  public double Lat { get; set; }
-  public double Lng { get; set; }
+  public float Lat { get; set; }
+  public float Lng { get; set; }
 }
